IoT_ProjectWeatherForcast/requirements.txt
fastapi
uvicorn[standard]
influxdb-client
python-dotenv
pydantic
python-multipart>=0.0.6
groq>=0.9.0
python-jose[cryptography]
passlib[bcrypt]
firebase-admin
requests
google-genai

